move merge_messages to aligner
